Change renderer to GLES3
When the repo was started for Wildling gamejam,
it was my understanding they would prefer an
HTML5 export, and according to the internet,
GLES2 would be the better way to go so that
the game would be playable in the browser on
lower end equipment. Since I'm not planning
to submit anymore, I'm switching back to
GLES3 so I can have muh features.